Hash-base-method/requirement.txt
# requirements.txt
# ----------------------------
torch
pillow
torchvision
scikit-learn
datasketch
matplotlib
tqdm